feat(GNB): /dashboard/[id] 페이지 GNB 적용
