+++ src/constants/router.ts
@@ -4,3 +4,5 @@ export const ROUTES = {
   DASHBOARD: '/dashboard',
   MY_DASHBOARD: '/mydashboard',
 };
+
+export const BASE_URL = 'https://sp-taskify-api.vercel.app/16-7';
