+++ src/constants/windowWidth.ts
@@ -1,0 +1,4 @@
+export const SMALL_DISPLAY = 768;
+export const BIG_DISPLAY = 1024;
+export const SMALLER_USERS_DISPLAY = 2;
+export const BIGGER_USERS_DISPLAY = 4;
